raspberrypi: Enable DSI display for rpi3
Set variable ENABLE_DSI to 1 in conf/local.conf to enable DSI in
Linux kernel 4.14 and get the official 7" Raspberry Pi touchscreen
display working with Raspberry Pi 3. By default ENABLE_DSI is set
to 0 and the image boots from HDMI.

Variable ENABLE_DSI is relevant for Raspberry Pi 3 only. This
workaround is NOT needed for Raspberry Pi 4.

On Raspberry Pi 4 both HDMI and the official 7" Raspberry Pi
touchscreen display work with Linux kernel 4.19. However for
Raspberry Pi 4 the DSI is with higher priority and the HMI will be
launched on it if it has been connected.

Bug-AGL: SPEC-2465

Signed-off-by: Leon Anavi <leon.anavi@konsulko.com>
Change-Id: Id967c57c9a316834f0bc23c6adedc1f7034fa1bd